+++ server/models/GenreModel.js
@@ -1,0 +1,9 @@
+import mongoose from "mongoose"
+
+const genreSchema = new mongoose.Schema({
+    name: { type: String, required: true}
+})
+
+const GenreModel = mongoose.model('Genre', genreSchema)
+
+export { GenreModel }
