+++ server/models/PlatformModel.js
@@ -1,0 +1,9 @@
+import mongoose from "mongoose"
+
+const platformSchema = new mongoose.Schema({
+    name: { type: String, required: true }
+})
+
+const PlatformModel = mongoose.model('Platform', platformSchema)
+
+export { PlatformModel }
